Column::make: return type static
`LinkColumn::make()->title()` gives currently an error in PHPStan in my project because it expects `LinkColumn::make` to return a `Column` even though it returns a `LinkColumn`.

I saw in your composer.json file that you also maintain compatability with 7.4, so I only changed the return type in the doc block. Ideal would of course be to also have `static` as return type also in the code itself.